tests/requirements.txt
pytest~=8.0.1
allure-pytest==2.13.2
selene>=2.0.0rc4
Appium-Python-Client~=3.1.1
python-dotenv==1.0.0
pydantic==2.5.2
requests~=2.31.0
